# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: BoardSync API
  version: 1.0.0
  description: |
    Kanban board with two-way Jira synchronisation. This file is the single source of truth
    for the REST contract (spec-first); the WebSocket contract is in `ws-events.md`.

    Conventions
    - Every operation requires a bearer JWT unless it overrides `security: []`.
    - Operations carrying `x-implemented-in: T-<n>` are part of the contract but are built by
      that ticket; operations without it are implemented.
    - Errors always use the `Error` schema.
    - Unbounded lists use cursor pagination (`limit`, `cursor`) and return `{ items, nextCursor }`;
      small board-scoped lists return plain arrays.
servers:
  - url: http://localhost:3000/api/v1
    description: Local backend
  - url: http://localhost:8080/api/v1
    description: Full compose stack (through the frontend proxy)
security:
  - bearerAuth: []
...
paths:
...
  /boards/{boardId}/columns:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    get:
      tags: [columns]
      operationId: listColumns
      summary: Columns of a board ordered by position
      responses:
        '200':
          description: Columns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Column'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [columns]
      operationId: createColumn
      summary: Add a column (Admin); appended unless a position is given
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColumnRequest'
      responses:
        '201':
          description: Created column
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Column'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /columns/{columnId}:
    parameters:
      - $ref: '#/components/parameters/ColumnId'
    patch:
      tags: [columns]
      operationId: updateColumn
      summary: Rename or retype a column (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateColumnRequest'
      responses:
        '200':
          description: Updated column
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Column'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [columns]
      operationId: deleteColumn
      summary: Delete an empty column while at least three remain (Admin)
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /columns/{columnId}/reorder:
    parameters:
      - $ref: '#/components/parameters/ColumnId'
    patch:
      tags: [columns]
      operationId: reorderColumn
      summary: Move a column to a position (Admin, transactional)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderColumnRequest'
      responses:
        '200':
          description: Columns of the board in their new order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Column'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/auth/register` or `/auth/login`, valid for at most one hour.
  parameters:
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ColumnId:
      name: columnId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation failed or the request is not acceptable in the current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing, invalid or expired token, or wrong credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The user's board role does not allow this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The resource does not exist or the user is not a member of its board
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The request conflicts with the current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: A message, or one message per invalid field
        details:
          type: object
          additionalProperties: true
...
    ColumnType:
      type: string
      enum: [TODO, IN_PROGRESS, REVIEW, DONE]
...
    Column:
      type: object
      required: [id, boardId, title, type, position]
      properties:
        id:
          type: string
          format: uuid
        boardId:
          type: string
          format: uuid
        title:
          type: string
        type:
          $ref: '#/components/schemas/ColumnType'
        position:
          type: integer
          minimum: 0
...
    CreateColumnRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        type:
          $ref: '#/components/schemas/ColumnType'
        position:
          type: integer
          minimum: 0
    UpdateColumnRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        type:
          $ref: '#/components/schemas/ColumnType'
    ReorderColumnRequest:
      type: object
      required: [position]
      properties:
        position:
          type: integer
          minimum: 0
...